List all nurseries on index page by default
The index page for nurseries has a group of checkboxes at the top of
the page to filter the nurseries displayed by district. All checkboxes
were unchecked by default and an empty list was displayed to the user.

Change the default so that all checkboxes are initially checked an thus
all nurseries are listed. The user can than narrow down the list by
unchecking certain districts.

Note: We can change `@nurseries` in the index action to `Nursery.all`
because the JavaScript form for filtering is send to the `results`
action, not the `index` action.